+++ src/base/SubscriptionGateway.php
@@ -67,7 +67,7 @@ abstract class SubscriptionGateway extends Gateway
      * });
      * ```
      */
-    const EVENT_CREATE_INVOICE = 'createInvoice';
+    public const EVENT_CREATE_INVOICE = 'createInvoice';
 
     /**
      * @event SubscriptionRequestEvent The event that is triggered when a subscription request is being built.
@@ -85,12 +85,12 @@ abstract class SubscriptionGateway extends Gateway
      * });
      * ```
      */
-    const EVENT_BEFORE_SUBSCRIBE = 'beforeSubscribe';
+    public const EVENT_BEFORE_SUBSCRIBE = 'beforeSubscribe';
 
     /**
      * string The Stripe API version to use.
      */
-    const STRIPE_API_VERSION = '2019-03-14';
+    public const STRIPE_API_VERSION = '2019-03-14';
 
     /**
      * @inheritdoc
@@ -172,7 +172,7 @@ abstract class SubscriptionGateway extends Gateway
     /**
      * @inheritdoc
      */
-    public function getPlanSettingsHtml(array $params = [])
+    public function getPlanSettingsHtml(array $params = []): ?string
     {
         $this->configureStripeClient();
         return Craft::$app->getView()->renderTemplate('commerce-stripe/planSettings', $params);
@@ -220,7 +220,7 @@ abstract class SubscriptionGateway extends Gateway
     /**
      * @inheritdoc
      */
-    public function refreshPaymentHistory(Subscription $subscription)
+    public function refreshpaymenthistory(Subscription $subscription): void
     {
         $this->configureStripeClient();
         // Update the subscription period.
@@ -457,6 +457,8 @@ abstract class SubscriptionGateway extends Gateway
      * @param Subscription $subscription
      * @param BasePlan $plan
      * @return float
+     * @throws \Stripe\Exception\ApiErrorException
+     * @throws \yii\base\InvalidConfigException
      */
     public function previewSwitchCost(Subscription $subscription, BasePlan $plan): float
     {
@@ -487,7 +489,7 @@ abstract class SubscriptionGateway extends Gateway
     /**
      * @inheritdoc
      */
-    public function handleWebhook(array $data)
+    public function handleWebhook(array $data): void
     {
         $this->configureStripeClient();
         switch ($data['type']) {
@@ -608,8 +610,9 @@ abstract class SubscriptionGateway extends Gateway
      * Handle a created invoice.
      *
      * @param array $data
+     * @throws \Stripe\Exception\ApiErrorException
      */
-    protected function handleInvoiceCreated(array $data)
+    protected function handleInvoiceCreated(array $data): void
     {
         $this->configureStripeClient();
         $stripeInvoice = $data['data']['object'];
@@ -634,7 +637,7 @@ abstract class SubscriptionGateway extends Gateway
      * @param array $data
      * @throws Throwable if something went wrong when processing the invoice
      */
-    protected function handleInvoiceSucceededEvent(array $data)
+    protected function handleInvoiceSucceededEvent(array $data): void
     {
         $this->configureStripeClient();
         $stripeInvoice = $data['data']['object'];
@@ -676,7 +679,7 @@ abstract class SubscriptionGateway extends Gateway
      * @param array $data
      * @throws InvalidConfigException If plan not available
      */
-    protected function handlePlanEvent(array $data)
+    protected function handlePlanEvent(array $data): void
     {
         $this->configureStripeClient();
         $planService = Commerce::getInstance()->getPlans();
@@ -698,7 +701,7 @@ abstract class SubscriptionGateway extends Gateway
      *
      * @throws Throwable
      */
-    protected function handleSubscriptionExpired(array $data)
+    protected function handleSubscriptionExpired(array $data): void
     {
         $this->configureStripeClient();
         $stripeSubscription = $data['data']['object'];
@@ -721,11 +724,11 @@ abstract class SubscriptionGateway extends Gateway
      *
      * @throws Throwable
      */
-    protected function handleSubscriptionUpdated(array $data)
+    protected function handleSubscriptionUpdated(array $data): void
     {
         $this->configureStripeClient();
         $stripeSubscription = $data['data']['object'];
-        $subscription = Subscription::find()->anyStatus()->reference($stripeSubscription['id'])->one();
+        $subscription = Subscription::find()->status(null)->reference($stripeSubscription['id'])->one();
 
         if (!$subscription) {
             Craft::warning('Subscription with the reference “' . $stripeSubscription['id'] . '” not found when processing webhook ' . $data['id'], 'stripe');
@@ -758,8 +761,9 @@ abstract class SubscriptionGateway extends Gateway
      * Set the various status properties on a Subscription by the subscription data set on it.
      *
      * @param Subscription $subscription
+     * @throws \Exception
      */
-    protected function setSubscriptionStatusData(Subscription $subscription)
+    protected function setSubscriptionStatusData(Subscription $subscription): void
     {
         $this->configureStripeClient();
         $subscriptionData = $subscription->getSubscriptionData();
@@ -808,6 +812,7 @@ abstract class SubscriptionGateway extends Gateway
      * @param array $stripeInvoice
      * @param Subscription $subscription
      * @return Invoice
+     * @throws \yii\base\Exception
      */
     protected function saveSubscriptionInvoice(array $stripeInvoice, Subscription $subscription): Invoice
     {