+++ src/events/BuildGatewayRequestEvent.php
@@ -22,15 +22,15 @@ class BuildGatewayRequestEvent extends Event
      * @var array The metadata of the gateway request
      * @deprecated in 1.1 Use [[request]] instead.
      */
-    public $metadata;
+    public array $metadata;
 
     /**
      * @var Transaction The transaction being used as the base for request
      */
-    public $transaction;
+    public Transaction $transaction;
 
     /**
      * @var array The request being used
      */
-    public $request;
+    public array $request;
 }